requirements.txt
asgiref==3.7.2
Django==5.0.3
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
django-daraja==1.3.0
django-mpesa==2.0.14